Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-dotenv
python-multipart
requests
apscheduler

# Optionnel mais recommandé
python-jose
passlib
redis
celery